refactor(views): compose track columns from one shared set
