+++ crates/views/src/screens/library/playlists.rs
@@ -20,58 +20,65 @@ pub(super) enum PlaylistField {
     TrackCount,
 }
 
-pub(super) const COLUMNS: &[ColumnSpec<PlaylistField>] = &[
-    ColumnSpec {
-        field: PlaylistField::Index,
-        key: "index",
-        header: "column-index",
-        align: TextAlign::Center,
-        width: Width::Fixed(NUMBER),
-        anchored: true,
-        sortable: false,
-        hide_below: ALWAYS,
-    },
-    ColumnSpec {
-        field: PlaylistField::Cover,
-        key: "cover",
-        header: "",
-        align: TextAlign::Left,
-        width: Width::Thumb,
-        anchored: true,
-        sortable: false,
-        hide_below: ALWAYS,
-    },
-    ColumnSpec {
-        field: PlaylistField::Name,
-        key: "name",
-        header: "column-name",
-        align: TextAlign::Left,
-        width: Width::Fill(0.55),
-        anchored: false,
-        sortable: true,
-        hide_below: ALWAYS,
-    },
-    ColumnSpec {
-        field: PlaylistField::Owner,
-        key: "owner",
-        header: "column-owner",
-        align: TextAlign::Left,
-        width: Width::Fill(0.45),
-        anchored: false,
-        sortable: true,
-        hide_below: ROOMY,
-    },
-    ColumnSpec {
-        field: PlaylistField::TrackCount,
-        key: "tracks",
-        header: "column-tracks",
-        align: TextAlign::Right,
-        width: Width::Fixed(TRAILING),
-        anchored: false,
-        sortable: true,
-        hide_below: SNUG,
-    },
-];
+const COLUMN: ColumnSpec<PlaylistField> = ColumnSpec {
+    field: PlaylistField::Index,
+    key: "",
+    header: "",
+    align: TextAlign::Left,
+    width: Width::Fill(1.),
+    anchored: false,
+    sortable: true,
+    hide_below: ALWAYS,
+};
+
+const INDEX: ColumnSpec<PlaylistField> = ColumnSpec {
+    field: PlaylistField::Index,
+    key: "index",
+    header: "column-index",
+    align: TextAlign::Center,
+    width: Width::Fixed(NUMBER),
+    anchored: true,
+    sortable: false,
+    ..COLUMN
+};
+
+const COVER: ColumnSpec<PlaylistField> = ColumnSpec {
+    field: PlaylistField::Cover,
+    key: "cover",
+    width: Width::Thumb,
+    anchored: true,
+    sortable: false,
+    ..COLUMN
+};
+
+const NAME: ColumnSpec<PlaylistField> = ColumnSpec {
+    field: PlaylistField::Name,
+    key: "name",
+    header: "column-name",
+    width: Width::Fill(0.55),
+    ..COLUMN
+};
+
+const OWNER: ColumnSpec<PlaylistField> = ColumnSpec {
+    field: PlaylistField::Owner,
+    key: "owner",
+    header: "column-owner",
+    width: Width::Fill(0.45),
+    hide_below: ROOMY,
+    ..COLUMN
+};
+
+const TRACK_COUNT: ColumnSpec<PlaylistField> = ColumnSpec {
+    field: PlaylistField::TrackCount,
+    key: "tracks",
+    header: "column-tracks",
+    align: TextAlign::Right,
+    width: Width::Fixed(TRAILING),
+    hide_below: SNUG,
+    ..COLUMN
+};
+
+pub(super) const COLUMNS: &[ColumnSpec<PlaylistField>] = &[INDEX, COVER, NAME, OWNER, TRACK_COUNT];
 
 pub(super) struct PlaylistSource {
     library: Entity<Library>,
